Configure number layer
Setup number layer with numbers on the left hand and function keys on
the right. Numbers can be used with combination of shift key to get
symbols. There's also a backspace key on a thumb to easily remove
mistyped numbers.

Finally, we set switch to the ADJ (adjust layer) on the left pinky.