Full home-row-mod on left and +1-row-mod on right
Keep the shift key on pinkies to preserve the muscle memory of using
pinkies to up case letters.

Then on the left use the "standard" home-row-mod location with the LGUI
on the index finger.

On the right hand we shift the "home-row-mod" to the +1 row (above the
home-row), to keep the vim navigation smooth. Re-tapping key to keep the
j/k/l repeating is annoying. But still we keep the shift on the
semicolon.

Some of the special keys like LGUI on left thumb cluster and RALT on
right thumb cluster are duplicated. To ease the transition to the new
layout.